@@ -1,5 +1,5 @@
 name: groovenation_flutter
-description: A new Flutter project.
+description: GrooveNation Mobile App
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,55 +27,56 @@ dependencies:
   cupertino_icons: ^1.0.2
   intl: ^0.17.0
   flutter_image_compress: ^1.0.0
-  percent_indicator: "^2.1.6"
+  percent_indicator: ^3.4.0
   font_awesome_flutter: ^9.0.0
-  flutter_rating_bar: ^3.0.1+1
-  carousel_slider: ^2.2.1
-  qr_flutter: ^3.1.0
-  flare_flutter: ^2.0.6
+  flutter_rating_bar: ^4.0.0
+  carousel_slider: ^4.0.0
+  qr_flutter: ^4.0.0
+  flare_flutter: ^3.0.2
   pull_to_refresh: ^2.0.0
   cached_network_image: ^3.1.0
-  #optimized_cached_image: ^2.0.1
-  image_picker: ^0.7.4
-  page_transition: "^1.1.6"
-  equatable: ^1.2.5
+  image_picker: ^0.8.4+4
+  page_transition: ^2.0.4
+  equatable: ^2.0.3
   dio: ^4.0.0
   bloc: ^7.0.0
   flutter_bloc: ^7.0.1
   hydrated_bloc: ^7.0.1
   hive: ^2.0.4
   hive_flutter: ^1.1.0
-  #auto_animated: any
   maps_launcher: ^2.0.0
   path_provider: ^2.0.1
-  share_plus: ^2.1.4
-  timeago: '2.0.25'
+  share_plus: ^3.0.4
+  timeago: '3.1.0'
   shared_preferences: ^2.0.6
   location: ^4.3.0
-  flutter_facebook_login: ^3.0.0
-  google_sign_in: ^4.5.6
+  # flutter_facebook_login: ^3.0.0
+  flutter_facebook_auth: ^3.5.3
+  google_sign_in: ^5.2.1
   image_cropper: ^1.3.0
   flutter_paystack: ^1.0.4+1
   url_launcher: ^6.0.6
   random_string: ^2.1.0
   crypt: ^4.0.1
-  sentry_flutter: ^4.0.6
-  #image_editor_pro: ^1.1.8
-  video_trimmer: ^0.5.1
-  file_picker: ^3.0.2+2
+  sentry_flutter: ^6.0.1
+  video_trimmer: ^0.6.0
+  file_picker: ^4.2.0
   video_player: ^2.1.6
-  video_thumbnail: ^0.3.3
+  video_thumbnail: ^0.4.3
   visibility_detector: ^0.2.0
   cached_video_player: ^2.0.0-dev.1
   socket_io_client: ^1.0.1
   firebase_core: "^1.4.0"
   firebase_messaging: "^10.0.4"
-  flutter_uploader: ^1.2.1
-  awesome_notifications: ^0.0.6+9
+  flutter_uploader: ^3.0.0-beta.3
+  awesome_notifications: ^0.6.13
   need_resume: ^1.0.6
   video_compress: ^3.1.0
   photofilters: ^3.0.1
-  # flutter_local_notifications: ^6.0.0
+  collection: ^1.15.0-nullsafety.4
+  firebase_analytics: ^9.0.0-dev.3
+  google_mobile_ads: ^0.13.6
+  # light_compressor: ^1.2.1
 
 dev_dependencies:
   flutter_test:
@@ -83,6 +84,7 @@ dev_dependencies:
 
   hive_generator: ^1.1.0
   build_runner: ^2.0.6
+  flutter_native_splash: ^1.3.1
   
 
 # For information on the generic Dart part of this file, see the
@@ -139,11 +141,6 @@ flutter:
     - family: LatoBlack
       fonts:
         - asset: fonts/Lato-Black.ttf
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
+
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
