@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:groovenation_flutter/constants/error.dart';
+import 'package:groovenation_flutter/constants/enums.dart';
 import 'package:groovenation_flutter/models/club.dart';
 
 abstract class ClubsState extends Equatable {}
@@ -10,14 +10,14 @@ class ClubsInitialState extends ClubsState {
 }
 
 class ClubsLoadingState extends ClubsState {
-  final List<Club> oldClubs;
+  final List<Club?>? oldClubs;
 
   ClubsLoadingState({
     this.oldClubs
   });
 
   @override
-  List<Object> get props => [oldClubs];
+  List<Object?> get props => [oldClubs];
 }
 
 class ClubFavouriteUpdatingState extends ClubsState {
@@ -26,8 +26,8 @@ class ClubFavouriteUpdatingState extends ClubsState {
 }
 
 class ClubsLoadedState extends ClubsState {
-  final List<Club> clubs;
-  final bool hasReachedMax;
+  final List<Club?>? clubs;
+  final bool? hasReachedMax;
 
   ClubsLoadedState({
     this.clubs,
@@ -35,25 +35,25 @@ class ClubsLoadedState extends ClubsState {
   });
 
   @override
-  List<Object> get props => [clubs, hasReachedMax];
+  List<Object?> get props => [clubs, hasReachedMax];
 }
 
 class ClubsErrorState extends ClubsState {
   ClubsErrorState(this.error);
 
-  final Error error;
+  final AppError error;
 
   @override
   List<Object> get props => [error];
 }
 
 class ClubLoadedState extends ClubsState {
-  final Club club;
+  final Club? club;
 
   ClubLoadedState({
     this.club,
   });
 
   @override
-  List<Object> get props => [club];
+  List<Object?> get props => [club];
 }
