@@ -1,13 +1,14 @@
 import 'package:crypt/crypt.dart';
 import 'package:dio/dio.dart';
-import 'package:groovenation_flutter/constants/auth_error_types.dart';
-import 'package:groovenation_flutter/constants/error.dart';
+import 'package:groovenation_flutter/constants/enums.dart';
 import 'package:groovenation_flutter/constants/strings.dart';
 import 'package:groovenation_flutter/models/auth_user.dart';
+import 'package:groovenation_flutter/models/city.dart';
+import 'package:groovenation_flutter/util/network_util.dart';
 import 'package:groovenation_flutter/util/shared_prefs.dart';
 
 class AuthRepository {
-  Future<AuthUser> loginUser(String email, String password) async {
+  Future<AuthUser?> loginUser(String email, String password) async {
     final c1 = Crypt.sha256(password, rounds: 10000, salt: PASSWORD_SHA_SALT);
 
     Map<String, String> json = {
@@ -18,9 +19,9 @@ class AuthRepository {
     return authenticateUserLogin("$API_HOST/users/login/email", json);
   }
 
-  Future<AuthUser> loginFacebook(
-      String email, String name, String facebookId) async {
-    Map<String, String> json = {
+  Future<AuthUser?> loginFacebook(
+      String? email, String? name, String? facebookId) async {
+    Map<String, String?> json = {
       "email": email,
       "name": name,
       "facebookId": facebookId,
@@ -29,9 +30,9 @@ class AuthRepository {
     return authenticateUserLogin("$API_HOST/users/login/facebook", json);
   }
 
-  Future<AuthUser> loginGoogle(
-      String email, String name, String googleId) async {
-    Map<String, String> json = {
+  Future<AuthUser?> loginGoogle(
+      String email, String? name, String googleId) async {
+    Map<String, String?> json = {
       "email": email,
       "name": name,
       "googleId": googleId,
@@ -40,7 +41,7 @@ class AuthRepository {
     return authenticateUserLogin("$API_HOST/users/login/google", json);
   }
 
-  Future<AuthUser> signup(String email, String firstName, String lastName,
+  Future<AuthUser?> signup(String email, String firstName, String lastName,
       String username, String password, DateTime dateOfBirth) async {
     final c1 = Crypt.sha256(password, rounds: 10000, salt: PASSWORD_SHA_SALT);
 
@@ -56,151 +57,106 @@ class AuthRepository {
     return authenticateUserSignup("$API_HOST/users/create/email", json);
   }
 
-  CancelToken _checkCancelToken;
-  Future<bool> checkUsernameExists(String username) async {
-    if (_checkCancelToken != null) {
-      try {
-        _checkCancelToken.cancel();
-        _checkCancelToken = null;
-      } catch (_) {}
-    }
+  CancelToken? _checkCancelToken;
+  Future<bool?> checkUsernameExists(String username) async {
+    NetworkUtil.cancel(_checkCancelToken);
 
+    _checkCancelToken = null;
     _checkCancelToken = CancelToken();
 
-    try {
-      Response response = await Dio().post("$API_HOST/users/check/username",
-          data: {
-            "username": username,
-          },
-          options: Options(contentType: Headers.formUrlEncodedContentType),
-          cancelToken: _checkCancelToken);
+    String url = "$API_HOST/users/check/username";
+    var body = {
+      "username": username,
+    };
 
-      if (response.statusCode == 200) {
-        Map<String, dynamic> jsonResponse = response.data;
+    var jsonResponse = await NetworkUtil.executePostRequest(
+        url, body, _onRequestError, _checkCancelToken);
 
-        if (jsonResponse['status'] == 1) {
-          print("uname - " +
-              username +
-              " - " +
-              jsonResponse['username_exists'].toString());
-          return jsonResponse['username_exists'];
-        } else
-          throw AuthException(Error.UNKNOWN_ERROR);
-      } else
-        throw AuthException(Error.UNKNOWN_ERROR);
-    } catch (e) {
-      if (e is AuthException)
-        throw AuthException(e.error);
-      else {
-        if (e is DioError) 
-        if (e.type == DioErrorType.cancel) {
-        } else
-          throw AuthException(Error.NETWORK_ERROR);
-        else
-          throw AuthException(Error.NETWORK_ERROR);
-      }
+    if (jsonResponse != null) {
+      return jsonResponse['username_exists'];
     }
+
+    return false;
   }
 
-  Future<AuthUser> authenticateUserLogin(
-      String url, Map<String, String> body) async {
-    try {
-      Response response = await Dio().post(url,
-          data: body,
-          options: Options(contentType: Headers.formUrlEncodedContentType));
+  Future<AuthUser?> authenticateUserLogin(
+      String url, Map<String, String?> body) async {
+    var jsonResponse =
+        await NetworkUtil.executePostRequest(url, body, _onRequestError);
 
-      print("Resp:" + response.data.toString());
-
-      if (response.statusCode == 200) {
-        Map<String, dynamic> jsonResponse = response.data;
-
-        if (jsonResponse['status'] == 1) {
-          if (jsonResponse['result'] == LOGIN_FAILED)
-            throw AuthLoginException(AuthLoginErrorType.LOGIN_FAILED);
-          else
-            return AuthUser.fromJson(jsonResponse['result']);
-        } else
-          throw AuthLoginException(AuthLoginErrorType.UNKNOWN_ERROR);
-      } else
-        throw AuthLoginException(AuthLoginErrorType.UNKNOWN_ERROR);
-    } catch (e) {
-      print(e);
-      if (e is AuthLoginException)
-        throw AuthLoginException(e.errorType);
+    if (jsonResponse != null) {
+      if (jsonResponse['result'] == LOGIN_FAILED)
+        throw AuthException(AuthError.LOGIN_FAILED);
       else
-        throw AuthLoginException(AuthLoginErrorType.UNKNOWN_ERROR);
+        return AuthUser.fromJson(jsonResponse['result']);
     }
+
+    return null;
   }
 
-  Future<AuthUser> authenticateUserSignup(
+  Future<AuthUser?> authenticateUserSignup(
       String url, Map<String, dynamic> body) async {
-    try {
-      Response response = await Dio().post(url,
-          data: body,
-          options: Options(contentType: Headers.formUrlEncodedContentType));
+    var jsonResponse =
+        await NetworkUtil.executePostRequest(url, body, _onRequestError);
 
-      if (response.statusCode == 200) {
-        Map<String, dynamic> jsonResponse = response.data;
-
-        if (jsonResponse['status'] == 1) {
-          if (jsonResponse['result'] == EMAIL_EXISTS)
-            throw AuthSignUpException(AuthSignUpErrorType.EMAIL_EXISTS_ERROR);
-          else if (jsonResponse['result'] == USERNAME_EXISTS)
-            throw AuthSignUpException(
-                AuthSignUpErrorType.USERNAME_EXISTS_ERROR);
-          else
-            return AuthUser.fromJson(jsonResponse['result']);
-        } else
-          throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
-      } else
-        throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
-    } catch (e) {
-      if (e is AuthSignUpException)
-        throw AuthSignUpException(e.errorType);
+    if (jsonResponse != null) {
+      if (jsonResponse['result'] == EMAIL_EXISTS)
+        throw AuthException(AuthError.EMAIL_EXISTS_ERROR);
+      else if (jsonResponse['result'] == USERNAME_EXISTS)
+        throw AuthException(AuthError.USERNAME_EXISTS_ERROR);
       else
-        throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
+        return AuthUser.fromJson(jsonResponse['result']);
     }
+
+    return null;
   }
 
-  Future<bool> createUsername(String username) async {
-    try {
-      Response response = await Dio().post("$API_HOST/users/create/username",
-          data: {'userId': sharedPrefs.userId, 'username': username},
-          options: Options(contentType: Headers.formUrlEncodedContentType));
+  Future<bool?> createUsername(String username) async {
+    String url = "$API_HOST/users/create/username";
+    var body = {'userId': sharedPrefs.userId, 'username': username};
 
-      if (response.statusCode == 200) {
-        Map<String, dynamic> jsonResponse = response.data;
+    var jsonResponse =
+        await NetworkUtil.executePostRequest(url, body, _onRequestError);
 
-        if (jsonResponse['status'] == 1) {
-          if (jsonResponse['result'] == USERNAME_EXISTS)
-            throw AuthSignUpException(
-                AuthSignUpErrorType.USERNAME_EXISTS_ERROR);
-          else
-            return jsonResponse['result'];
-        } else
-          throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
-      } else
-        throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
-    } catch (e) {
-      if (e is AuthSignUpException)
-        throw AuthSignUpException(e.errorType);
+    if (jsonResponse != null) {
+      if (jsonResponse['result'] == USERNAME_EXISTS)
+        throw AuthException(AuthError.USERNAME_EXISTS_ERROR);
       else
-        throw AuthSignUpException(AuthSignUpErrorType.UNKNOWN_ERROR);
+        return jsonResponse['result'];
     }
+
+    return null;
   }
-}
 
-class AuthSignUpException implements Exception {
-  final AuthSignUpErrorType errorType;
-  AuthSignUpException(this.errorType);
-}
+  Future<List<City>?> getAvailableCities() async {
+    List<City> cities = [];
 
-class AuthLoginException implements Exception {
-  final AuthLoginErrorType errorType;
-  AuthLoginException(this.errorType);
+    var jsonResponse = await NetworkUtil.executeGetRequest(
+        "$API_HOST/users/cities/available", _onRequestError);
+
+    if (jsonResponse != null) {
+      for (Map i in jsonResponse['cities']) {
+        City city = City.fromJson(i);
+        cities.add(city);
+      }
+
+      return cities;
+    }
+
+    return null;
+  }
+
+  _onRequestError(e) {
+    if (e is AuthException)
+      throw AuthException(e.error);
+    else if (e is DioError) if (e.type != DioErrorType.cancel)
+      throw AuthException(AuthError.NETWORK_ERROR);
+    else
+      throw AuthException(AuthError.UNKNOWN_ERROR);
+  }
 }
 
 class AuthException implements Exception {
-  final Error error;
+  final AuthError error;
   AuthException(this.error);
 }
