@@ -1,6 +1,6 @@
 class APIResult {
   final dynamic result;
-  final bool hasReachedMax;
+  final bool? hasReachedMax;
 
   APIResult(this.result, this.hasReachedMax);
 }
