@@ -1,12 +1,13 @@
 class AuthUser {
-  final String userId;
-  final String username;
-  final String userCity;
-  final String profilePicUrl;
-  final String coverPicUrl;
+  final String? userId;
+  final String? username;
+  final String? userCity;
+  final String? profilePicUrl;
+  final String? coverPicUrl;
+  final String? authToken;
 
   AuthUser(this.userId, this.username, this.userCity, this.profilePicUrl,
-      this.coverPicUrl);
+      this.coverPicUrl, this.authToken);
 
   factory AuthUser.fromJson(dynamic json) {
     return AuthUser(
@@ -15,6 +16,7 @@ class AuthUser {
       json['userCity'],
       json['profilePicUrl'],
       json['coverPicUrl'],
+      json['auth_token'],
     );
   }
 }
