@@ -8,25 +8,25 @@ part 'message.g.dart';
 @HiveType(typeId: 1)
 class Message {
   @HiveField(0)
-  final String messageID;
+  final String? messageID;
 
   @HiveField(1)
-  String conversationId;
+  String? conversationId;
 
   @HiveField(3)
-  final String messageType;
+  final String? messageType;
 
   @HiveField(4)
-  final DateTime messageDateTime;
+  final DateTime? messageDateTime;
 
   @HiveField(5)
-  final SocialPerson sender;
+  final SocialPerson? sender;
 
   @HiveField(6)
-  final String receiverId;
+  final String? receiverId;
   
   @HiveField(7)
-  String messageStatus;
+  String? messageStatus;
 
   Message(this.messageID, this.conversationId, this.messageType,
       this.messageDateTime, this.sender, this.receiverId, [this.messageStatus = MESSAGE_STATUS_PENDING]);
@@ -45,19 +45,19 @@ class Message {
   }
 
   static Map toJson(Message m) {
-    if(m.messageType == MESSAGE_TYPE_MEDIA) return MediaMessage.getJson(m); 
-    else if(m.messageType == MESSAGE_TYPE_POST) return SocialPostMessage.getJson(m);
-    else return TextMessage.getJson(m);
+    if(m.messageType == MESSAGE_TYPE_MEDIA) return MediaMessage.getJson(m as MediaMessage); 
+    else if(m.messageType == MESSAGE_TYPE_POST) return SocialPostMessage.getJson(m as SocialPostMessage);
+    else return TextMessage.getJson(m as TextMessage);
   }
 }
 
 @HiveType(typeId: 3)
 class TextMessage extends Message {
   @HiveField(8)
-  final String text;
+  final String? text;
 
-  TextMessage(String messageID, String conversationId, DateTime messageDateTime,
-      SocialPerson sender, this.text, String receiverId)
+  TextMessage(String? messageID, String? conversationId, DateTime? messageDateTime,
+      SocialPerson? sender, this.text, String? receiverId)
       : super(messageID, conversationId, MESSAGE_TYPE_TEXT, messageDateTime,
             sender, receiverId);
 
@@ -76,8 +76,8 @@ class TextMessage extends Message {
       "messageId" : message.messageID,
       "conversationId" : message.conversationId,
       "messageType" : message.messageType,
-      "messageDateTime" : message.messageDateTime.toIso8601String(),
-      "sender" : message.sender.toJson(),
+      "messageDateTime" : message.messageDateTime!.toIso8601String(),
+      "sender" : message.sender!.toJson(),
       "text" : message.text,
       "receiverId" : message.receiverId,
     };
@@ -87,15 +87,15 @@ class TextMessage extends Message {
 @HiveType(typeId: 4)
 class MediaMessage extends Message {
   @HiveField(8)
-  final String mediaURL;
+  final String? mediaURL;
 
   MediaMessage(
-      String messageID,
-      String conversationId,
-      DateTime messageDateTime,
-      SocialPerson sender,
+      String? messageID,
+      String? conversationId,
+      DateTime? messageDateTime,
+      SocialPerson? sender,
       this.mediaURL,
-      String receiverId, [String messageStatus])
+      String? receiverId, [String? messageStatus])
       : super(messageID, conversationId, MESSAGE_TYPE_MEDIA, messageDateTime,
             sender, receiverId, messageStatus);
 
@@ -114,8 +114,8 @@ class MediaMessage extends Message {
       "messageId" : message.messageID,
       "conversationId" : message.conversationId,
       "messageType" : message.messageType,
-      "messageDateTime" : message.messageDateTime.toIso8601String(),
-      "sender" : message.sender.toJson(),
+      "messageDateTime" : message.messageDateTime!.toIso8601String(),
+      "sender" : message.sender!.toJson(),
       "mediaURL" : message.mediaURL,
       "receiverId" : message.receiverId,
     };
@@ -125,15 +125,15 @@ class MediaMessage extends Message {
 @HiveType(typeId: 5)
 class SocialPostMessage extends Message {
   @HiveField(8)
-  final SocialPost post;
+  final SocialPost? post;
 
   SocialPostMessage(
-      String messageID,
-      String conversationId,
-      DateTime messageDateTime,
-      SocialPerson sender,
+      String? messageID,
+      String? conversationId,
+      DateTime? messageDateTime,
+      SocialPerson? sender,
       this.post,
-      String receiverId)
+      String? receiverId)
       : super(messageID, conversationId, MESSAGE_TYPE_POST, messageDateTime,
             sender, receiverId);
 
@@ -152,9 +152,9 @@ class SocialPostMessage extends Message {
       "messageId" : message.messageID,
       "conversationId" : message.conversationId,
       "messageType" : message.messageType,
-      "messageDateTime" : message.messageDateTime.toIso8601String(),
-      "sender" : message.sender.toJson(),
-      "post" : message.post.toJson(),
+      "messageDateTime" : message.messageDateTime!.toIso8601String(),
+      "sender" : message.sender!.toJson(),
+      "post" : message.post!.toJson(),
       "receiverId" : message.receiverId,
     };
   }
