@@ -5,10 +5,10 @@ part 'saved_message.g.dart';
 @HiveType(typeId: 6)
 class SavedMessage {
   @HiveField(0)
-  String conversationId;
+  String? conversationId;
 
   @HiveField(1)
-  Map messageJSON;
+  Map? messageJSON;
 
   SavedMessage(this.conversationId, this.messageJSON);
 }
