@@ -5,13 +5,13 @@ part 'send_media_task.g.dart';
 @HiveType(typeId: 7)
 class SendMediaTask {
   @HiveField(0)
-  String taskId;
+  String? taskId;
 
   @HiveField(1)
-  String filePath;
+  String? filePath;
   
   @HiveField(2)
-  String receiverId;
+  String? receiverId;
 
   SendMediaTask(this.taskId, this.filePath, this.receiverId);
 }
