@@ -2,18 +2,19 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:groovenation_flutter/constants/error.dart';
+import 'package:groovenation_flutter/constants/enums.dart';
 import 'package:groovenation_flutter/constants/strings.dart';
-import 'package:groovenation_flutter/cubit/social_cubit.dart';
+import 'package:groovenation_flutter/cubit/club/club_moments_cubit.dart';
 import 'package:groovenation_flutter/cubit/state/social_state.dart';
 import 'package:groovenation_flutter/models/club.dart';
 import 'package:groovenation_flutter/models/social_post.dart';
-import 'package:groovenation_flutter/ui/social/social_item.dart';
+import 'package:groovenation_flutter/ui/social/widgets/social_item.dart';
 import 'package:groovenation_flutter/util/alert_util.dart';
+import 'package:groovenation_flutter/widgets/custom_refresh_header.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class ClubMomentsPage extends StatefulWidget {
-  final Club club;
+  final Club? club;
   ClubMomentsPage(this.club);
 
   @override
@@ -21,7 +22,7 @@ class ClubMomentsPage extends StatefulWidget {
 }
 
 class _ClubMomentsPageState extends State<ClubMomentsPage> {
-  final Club club;
+  final Club? club;
   _ClubMomentsPageState(this.club);
 
   bool _scrollToTopVisible = false;
@@ -48,7 +49,7 @@ class _ClubMomentsPageState extends State<ClubMomentsPage> {
     });
     final ClubMomentsCubit clubSocialCubit =
         BlocProvider.of<ClubMomentsCubit>(context);
-    clubSocialCubit.getMoments(_page, club.clubID);
+    clubSocialCubit.getMoments(_page, club!.clubID);
   }
 
   @override
@@ -57,8 +58,8 @@ class _ClubMomentsPageState extends State<ClubMomentsPage> {
     super.dispose();
   }
 
-  List<SocialPost> clubSocial = [];
-  bool hasReachedMax = false;
+  List<SocialPost?>? clubSocial = [];
+  bool? hasReachedMax = false;
   int _page = 0;
 
   _onRefresh() {
@@ -67,18 +68,18 @@ class _ClubMomentsPageState extends State<ClubMomentsPage> {
     final ClubMomentsCubit clubSocialCubit =
         BlocProvider.of<ClubMomentsCubit>(context);
     if (!(clubSocialCubit.state is SocialLoadingState)) {
-      clubSocialCubit.getMoments(0, club.clubID);
+      clubSocialCubit.getMoments(0, club!.clubID);
     }
   }
 
   _onLoading() {
-    if (clubSocial.length == 0 || hasReachedMax) return;
+    if (clubSocial!.length == 0 || hasReachedMax!) return;
 
     _page++;
     final ClubMomentsCubit clubSocialCubit =
         BlocProvider.of<ClubMomentsCubit>(context);
     if (!(clubSocialCubit.state is SocialLoadingState)) {
-      clubSocialCubit.getMoments(_page, club.clubID);
+      clubSocialCubit.getMoments(_page, club!.clubID);
     }
   }
 
@@ -97,165 +98,174 @@ class _ClubMomentsPageState extends State<ClubMomentsPage> {
     SystemChrome.setSystemUIOverlayStyle(myTheme);
 
     return SafeArea(
-        child: Stack(children: [
-      // CustomScrollView(
-      //   controller: _scrollController,
-      //   physics: const BouncingScrollPhysics(
-      //       parent: AlwaysScrollableScrollPhysics()),
-      //   slivers: [
-      //     SliverToBoxAdapter(
-      //       child:
-      Padding(
-          padding: EdgeInsets.only(top: 16),
-          child: Column(
-            children: [
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Padding(
-                      padding: EdgeInsets.only(left: 16, top: 16),
-                      child: Container(
-                        height: 48,
-                        width: 48,
-                        decoration: BoxDecoration(
-                            color: Colors.deepPurple,
-                            borderRadius: BorderRadius.circular(900)),
-                        child: FlatButton(
-                          padding: EdgeInsets.only(left: 8),
-                          onPressed: () {
-                            Navigator.pop(context);
-                          },
-                          child: Icon(
-                            Icons.arrow_back_ios,
-                            color: Colors.white,
-                            size: 24,
-                          ),
-                        ),
-                      )),
-                  Padding(
-                      padding: EdgeInsets.only(left: 36, top: 16),
-                      child: Text(
-                        "Recent Moments",
-                        style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 36,
-                            fontFamily: 'LatoBold'),
-                      )),
-                ],
+      child: Stack(
+        children: [
+          Padding(
+            padding: EdgeInsets.only(top: 16),
+            child: Column(
+              children: [
+                _header(),
+                _momentsList(),
+              ],
+            ),
+          ),
+          _scrollToTopButton(),
+        ],
+      ),
+    );
+  }
+
+  Widget _header() {
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        Padding(
+          padding: EdgeInsets.only(left: 16, top: 16),
+          child: Container(
+            height: 48,
+            width: 48,
+            decoration: BoxDecoration(
+                color: Colors.deepPurple,
+                borderRadius: BorderRadius.circular(900)),
+            child: FlatButton(
+              padding: EdgeInsets.only(left: 8),
+              onPressed: () {
+                Navigator.pop(context);
+              },
+              child: Icon(
+                Icons.arrow_back_ios,
+                color: Colors.white,
+                size: 24,
               ),
-              Expanded(
-                  child: SmartRefresher(
-                      controller: _momentRefreshController,
-                      header: WaterDropMaterialHeader(),
-                      footer: _classicFooter,
-                      onLoading: () => _onLoading(),
-                      onRefresh: () => _onRefresh(),
-                      enablePullDown: true,
-                      enablePullUp: true,
-                      child: BlocConsumer<ClubMomentsCubit, SocialState>(
-                          listener: (context, state) {
-                        if (state is SocialLoadedState) {
-                          _momentRefreshController.refreshCompleted();
+            ),
+          ),
+        ),
+        Padding(
+          padding: EdgeInsets.only(left: 36, top: 16),
+          child: Text(
+            "Recent Moments",
+            style: TextStyle(
+                color: Colors.white, fontSize: 36, fontFamily: 'LatoBold'),
+          ),
+        ),
+      ],
+    );
+  }
 
-                          if (state.hasReachedMax)
-                            _momentRefreshController.loadNoData();
-                        }
-                        if (state is SocialErrorState) {
-                          switch (state.error) {
-                            case Error.NETWORK_ERROR:
-                              alertUtil.sendAlert(
-                                  BASIC_ERROR_TITLE,
-                                  NETWORK_ERROR_PROMPT,
-                                  Colors.red,
-                                  Icons.error);
-                              break;
-                            default:
-                              alertUtil.sendAlert(
-                                  BASIC_ERROR_TITLE,
-                                  UNKNOWN_ERROR_PROMPT,
-                                  Colors.red,
-                                  Icons.error);
-                              break;
-                          }
-                        }
-                      }, builder: (context, state) {
-                        if (state is SocialLoadingState) {
-                          return Padding(
-                              padding: EdgeInsets.only(top: 64),
-                              child: Center(
-                                  child: SizedBox(
-                                height: 56,
-                                width: 56,
-                                child: CircularProgressIndicator(
-                                  valueColor: AlwaysStoppedAnimation<Color>(
-                                      Colors.white),
-                                  strokeWidth: 2.0,
-                                ),
-                              )));
-                        }
+  void _blocListener(context, state) {
+    if (state is SocialLoadedState) {
+      _momentRefreshController.refreshCompleted();
 
-                        if (state is SocialLoadedState) {
-                          clubSocial = state.socialPosts;
-                          hasReachedMax = state.hasReachedMax;
-                        }
+      if (state.hasReachedMax!) _momentRefreshController.loadNoData();
+    }
+    if (state is SocialErrorState) {
+      switch (state.error) {
+        case AppError.NETWORK_ERROR:
+          alertUtil.sendAlert(
+              BASIC_ERROR_TITLE, NETWORK_ERROR_PROMPT, Colors.red, Icons.error);
+          break;
+        default:
+          alertUtil.sendAlert(
+              BASIC_ERROR_TITLE, UNKNOWN_ERROR_PROMPT, Colors.red, Icons.error);
+          break;
+      }
+    }
+  }
 
-                        return ListView.builder(
-                            padding: EdgeInsets.only(top: 12, bottom: 12),
-                            physics: NeverScrollableScrollPhysics(),
-                            shrinkWrap: true,
-                            itemCount: clubSocial.length,
-                            itemBuilder: (context, index) {
-                              return Padding(
-                                  padding: EdgeInsets.fromLTRB(12, 16, 12, 0),
-                                  child: Align(
-                                    child: SocialItem(
-                                      key: Key(clubSocial[index].postID),
-                                        socialPost: clubSocial[index],
-                                        showClose: false),
-                                    alignment: Alignment.topCenter,
-                                  ));
-                            });
-                      })))
-            ],
-          )),
-      //     )
-      //   ],
-      // ),
-      AnimatedOpacity(
-          opacity: _scrollToTopVisible ? 1.0 : 0.0,
-          duration: Duration(milliseconds: 250),
-          child: Align(
-              alignment: Alignment.bottomRight,
-              child: Padding(
-                  padding: EdgeInsets.only(bottom: 24, right: 24),
-                  child: Card(
-                    elevation: 6,
-                    color: Colors.deepPurple,
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(9)),
-                    child: Container(
-                      height: 48,
-                      width: 48,
-                      decoration: BoxDecoration(
-                          color: Colors.deepPurple,
-                          borderRadius: BorderRadius.circular(9)),
-                      child: FlatButton(
-                        padding: EdgeInsets.zero,
-                        onPressed: () {
-                          _scrollController.animateTo(
-                            0.0,
-                            curve: Curves.easeOut,
-                            duration: const Duration(milliseconds: 300),
-                          );
-                        },
-                        child: Icon(
-                          Icons.keyboard_arrow_up,
-                          color: Colors.white.withOpacity(0.8),
-                          size: 36,
-                        ),
-                      ),
+  Widget _momentsList() {
+    return Expanded(
+      child: SmartRefresher(
+        controller: _momentRefreshController,
+        header: CustomMaterialClassicHeader(),
+        footer: _classicFooter,
+        onLoading: () => _onLoading(),
+        onRefresh: () => _onRefresh(),
+        enablePullDown: true,
+        enablePullUp: true,
+        child: BlocConsumer<ClubMomentsCubit, SocialState>(
+          listener: _blocListener,
+          builder: (context, state) {
+            if (state is SocialLoadingState) {
+              return Padding(
+                  padding: EdgeInsets.only(top: 64),
+                  child: Center(
+                      child: SizedBox(
+                    height: 56,
+                    width: 56,
+                    child: CircularProgressIndicator(
+                      valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+                      strokeWidth: 2.0,
                     ),
-                  ))))
-    ]));
+                  )));
+            }
+
+            if (state is SocialLoadedState) {
+              clubSocial = state.socialPosts;
+              hasReachedMax = state.hasReachedMax;
+            }
+
+            return ListView.builder(
+              padding: EdgeInsets.only(top: 12, bottom: 12),
+              physics: NeverScrollableScrollPhysics(),
+              shrinkWrap: true,
+              itemCount: clubSocial!.length,
+              itemBuilder: (context, index) {
+                return Padding(
+                  padding: EdgeInsets.fromLTRB(12, 16, 12, 0),
+                  child: Align(
+                    child: SocialItem(
+                        key: Key(clubSocial![index]!.postID!),
+                        socialPost: clubSocial![index],
+                        showClose: false),
+                    alignment: Alignment.topCenter,
+                  ),
+                );
+              },
+            );
+          },
+        ),
+      ),
+    );
+  }
+
+  Widget _scrollToTopButton() {
+    return AnimatedOpacity(
+      opacity: _scrollToTopVisible ? 1.0 : 0.0,
+      duration: Duration(milliseconds: 250),
+      child: Align(
+        alignment: Alignment.bottomRight,
+        child: Padding(
+          padding: EdgeInsets.only(bottom: 24, right: 24),
+          child: Card(
+            elevation: 6,
+            color: Colors.deepPurple,
+            shape:
+                RoundedRectangleBorder(borderRadius: BorderRadius.circular(9)),
+            child: Container(
+              height: 48,
+              width: 48,
+              decoration: BoxDecoration(
+                  color: Colors.deepPurple,
+                  borderRadius: BorderRadius.circular(9)),
+              child: FlatButton(
+                padding: EdgeInsets.zero,
+                onPressed: () {
+                  _scrollController.animateTo(
+                    0.0,
+                    curve: Curves.easeOut,
+                    duration: const Duration(milliseconds: 300),
+                  );
+                },
+                child: Icon(
+                  Icons.keyboard_arrow_up,
+                  color: Colors.white.withOpacity(0.8),
+                  size: 36,
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
   }
 }
