@@ -3,7 +3,7 @@ import 'package:groovenation_flutter/models/message.dart';
 
 class ChatPageArguments{
   final Conversation conversation;
-  final Message messageToSend;
+  final Message? messageToSend;
 
   ChatPageArguments(this.conversation, this.messageToSend);
 }
