@@ -17,9 +17,9 @@ class SendMediaTaskAdapter extends TypeAdapter<SendMediaTask> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return SendMediaTask(
-      fields[0] as String,
-      fields[1] as String,
-      fields[2] as String,
+      fields[0] as String?,
+      fields[1] as String?,
+      fields[2] as String?,
     );
   }
 
