@@ -17,12 +17,12 @@ class SocialPersonAdapter extends TypeAdapter<SocialPerson> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return SocialPerson(
-      fields[0] as String,
-      fields[1] as String,
-      fields[2] as String,
-      fields[3] as String,
-      fields[4] as bool,
-      fields[5] as bool,
+      fields[0] as String?,
+      fields[1] as String?,
+      fields[2] as String?,
+      fields[3] as String?,
+      fields[4] as bool?,
+      fields[5] as bool?,
     );
   }
 
