@@ -17,20 +17,20 @@ class TicketAdapter extends TypeAdapter<Ticket> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Ticket(
-      fields[0] as String,
-      fields[1] as String,
-      fields[2] as String,
-      fields[3] as String,
-      fields[4] as String,
-      fields[5] as double,
-      fields[6] as double,
-      fields[7] as DateTime,
-      fields[8] as DateTime,
-      fields[9] as String,
-      fields[10] as int,
-      fields[11] as int,
-      fields[12] as String,
-      fields[13] as bool,
+      fields[0] as String?,
+      fields[1] as String?,
+      fields[2] as String?,
+      fields[3] as String?,
+      fields[4] as String?,
+      fields[5] as double?,
+      fields[6] as double?,
+      fields[7] as DateTime?,
+      fields[8] as DateTime?,
+      fields[9] as String?,
+      fields[10] as int?,
+      fields[11] as int?,
+      fields[12] as String?,
+      fields[13] as bool?,
     );
   }
 
