@@ -1,8 +1,10 @@
 import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:groovenation_flutter/constants/strings.dart';
-import 'package:groovenation_flutter/ui/screens/main_app_page.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:groovenation_flutter/cubit/state/auth_state.dart';
+import 'package:groovenation_flutter/cubit/user/auth_cubit.dart';
+import 'package:groovenation_flutter/models/city.dart';
 import 'package:groovenation_flutter/util/shared_prefs.dart';
 import 'package:location/location.dart';
 import 'package:cached_network_image/cached_network_image.dart';
@@ -13,6 +15,18 @@ class CityPickerPage extends StatefulWidget {
 }
 
 class _CityPickerPageState extends State<CityPickerPage> {
+  String? _selectedCity;
+  double? _selectedDefaultLat;
+  double? _selectedDefaultLon;
+
+  @override
+  void initState() {
+    super.initState();
+
+    AuthCubit authCubit = BlocProvider.of<AuthCubit>(context);
+    authCubit.getAvailableCities();
+  }
+
   Future<void> _showPermissionDialog() async {
     return showDialog<void>(
       context: context,
@@ -25,23 +39,23 @@ class _CityPickerPageState extends State<CityPickerPage> {
           ),
           content: SingleChildScrollView(
             child: ListBody(
-              children: <Widget>[
+              children: [
                 Text(
                     "Optional: GrooveNation can use your location to provide more accurate data on nearby clubs and events. Will you allow GrooveNation to access your location when available?",
                     style: TextStyle(fontFamily: 'Lato')),
               ],
             ),
           ),
-          actions: <Widget>[
+          actions: [
             FlatButton(
               child: Text("No"),
               onPressed: () {
-                sharedPrefs.userCity = selectedCity;
+                sharedPrefs.userCity = _selectedCity;
+                sharedPrefs.defaultLat = _selectedDefaultLat;
+                sharedPrefs.defaultLon = _selectedDefaultLon;
+
                 if (sharedPrefs.username != null)
-                  Navigator.pushReplacement(
-                      context,
-                      new MaterialPageRoute(
-                          builder: (context) => new MainNavigationScreen()));
+                  Navigator.pushReplacementNamed(context, '/main');
                 else
                   Navigator.pushReplacementNamed(context, '/create_username');
               },
@@ -52,12 +66,12 @@ class _CityPickerPageState extends State<CityPickerPage> {
                 Location location = new Location();
                 await location.requestPermission();
 
-                sharedPrefs.userCity = selectedCity;
+                sharedPrefs.userCity = _selectedCity;
+                sharedPrefs.defaultLat = _selectedDefaultLat;
+                sharedPrefs.defaultLon = _selectedDefaultLon;
+
                 if (sharedPrefs.username != null)
-                  Navigator.push(
-                      context,
-                      new MaterialPageRoute(
-                          builder: (context) => new MainNavigationScreen()));
+                  Navigator.pushReplacementNamed(context, '/main');
                 else
                   Navigator.pushReplacementNamed(context, '/create_username');
               },
@@ -68,44 +82,21 @@ class _CityPickerPageState extends State<CityPickerPage> {
     );
   }
 
-  Column topAppBar() => Column(children: [
-        Padding(
-          padding: EdgeInsets.only(top: 16, left: 16, right: 16),
-          child: Container(
-            width: double.infinity,
-            decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10.0),
-                color: Colors.white.withOpacity(0.2)),
-            child: TextField(
-              keyboardType: TextInputType.multiline,
-              autofocus: false,
-              cursorColor: Colors.white.withOpacity(0.7),
-              style: TextStyle(
-                  fontFamily: 'Lato', color: Colors.white, fontSize: 20),
-              decoration: InputDecoration(
-                  hintMaxLines: 3,
-                  border: InputBorder.none,
-                  focusedBorder: InputBorder.none,
-                  enabledBorder: InputBorder.none,
-                  errorBorder: InputBorder.none,
-                  disabledBorder: InputBorder.none,
-                  contentPadding:
-                      EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-                  hintText: "Search...",
-                  hintStyle: TextStyle(color: Colors.white.withOpacity(0.2)),
-                  suffixIcon: Padding(
-                      padding: EdgeInsets.only(right: 10),
-                      child: Icon(
-                        Icons.search,
-                        color: Colors.white.withOpacity(0.2),
-                        size: 28,
-                      ))),
-            ),
+  Widget _circularProgress() {
+    return Padding(
+      padding: EdgeInsets.only(top: 16),
+      child: Center(
+        child: SizedBox(
+          height: 56,
+          width: 56,
+          child: CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+            strokeWidth: 2.0,
           ),
         ),
-      ]);
-
-  String selectedCity;
+      ),
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -115,130 +106,164 @@ class _CityPickerPageState extends State<CityPickerPage> {
     SystemChrome.setSystemUIOverlayStyle(myTheme);
 
     return SafeArea(
-        child: CustomScrollView(
-      physics:
-          const BouncingScrollPhysics(parent: AlwaysScrollableScrollPhysics()),
-      slivers: [
-        SliverToBoxAdapter(
+      child: CustomScrollView(
+        physics: const BouncingScrollPhysics(
+            parent: AlwaysScrollableScrollPhysics()),
+        slivers: [
+          SliverToBoxAdapter(
             child: Padding(
-                padding: EdgeInsets.only(top: 16),
-                child: Column(
-                  children: [
-                    Padding(
-                        padding: EdgeInsets.only(left: 16, top: 8),
-                        child: Text(
-                          "Choose A City",
-                          style: TextStyle(
-                              color: Colors.white,
-                              fontSize: 32,
-                              fontFamily: 'LatoBold'),
-                        )),
-                    Visibility(
-                      child: topAppBar(),
-                      visible: false,
+              padding: EdgeInsets.only(top: 16),
+              child: Column(
+                children: [
+                  Padding(
+                    padding: EdgeInsets.only(top: 16),
+                    child: Text(
+                      "Choose A City",
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 28,
+                          fontFamily: 'LatoBold'),
                     ),
-                    ListView(
-                      padding: EdgeInsets.only(top: 12),
-                      physics: NeverScrollableScrollPhysics(),
-                      shrinkWrap: true,
-                      children: [
-                        Padding(
-                            padding: EdgeInsets.fromLTRB(12, 0, 12, 4),
+                  ),
+                  BlocBuilder<AuthCubit, AuthState>(builder: (context, state) {
+                    List<City> cities = [];
+
+                    if (state is AuthAvailableCitiesLoadingState) {
+                      return _circularProgress();
+                    } else if (state is AuthAvailableCitiesErrorState) {
+                      return Container(
+                        padding:
+                            EdgeInsets.symmetric(vertical: 48, horizontal: 24),
+                        child: Text(
+                          "Something Went Wrong. Please check your connection and try again",
+                          textAlign: TextAlign.center,
+                          style: TextStyle(
+                              color: Colors.white.withOpacity(0.5),
+                              fontSize: 16,
+                              fontFamily: 'Lato'),
+                        ),
+                      );
+                    }
+
+                    if (state is AuthAvailableCitiesLoadedState)
+                      cities = state.cities;
+
+                    return ListView.builder(
+                        padding: EdgeInsets.only(top: 16),
+                        physics: NeverScrollableScrollPhysics(),
+                        shrinkWrap: true,
+                        itemCount: cities.length,
+                        itemBuilder: (context, index) {
+                          return Padding(
+                            padding: EdgeInsets.fromLTRB(12, 0, 12, 0),
                             child: Align(
-                              child: cityItem(context, "Johannesburg",
-                                  "https://media.threatpost.com/wp-content/uploads/sites/103/2019/10/30083110/johannesburg-e1572438686227.jpg"),
+                              child: _cityItem(context, cities[index]),
                               alignment: Alignment.topCenter,
-                            )),
-                      ],
-                    )
-                  ],
-                )))
-      ],
-    ));
+                            ),
+                          );
+                        });
+                  }),
+                ],
+              ),
+            ),
+          )
+        ],
+      ),
+    );
   }
 
-  Widget cityItem(BuildContext context, String name, String imageUrl) {
+  Widget _cityItem(BuildContext context, City city) {
     return Padding(
-        padding: EdgeInsets.symmetric(vertical: 6),
-        child: Card(
-          elevation: 4,
-          color: Colors.deepPurple,
-          clipBehavior: Clip.antiAliasWithSaveLayer,
-          shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.all(Radius.circular(12.0))),
-          child: FlatButton(
-              onPressed: () {
-                selectedCity = CITY_JOHANNESBURG;
-                _showPermissionDialog();
-              },
-              padding: EdgeInsets.zero,
-              child: Wrap(children: [
-                Column(
-                  children: [
-                    Padding(
-                        padding: EdgeInsets.all(16),
-                        child: Column(
+      padding: EdgeInsets.symmetric(vertical: 6),
+      child: Card(
+        elevation: 4,
+        color: Colors.deepPurple,
+        clipBehavior: Clip.antiAliasWithSaveLayer,
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.all(Radius.circular(12.0))),
+        child: FlatButton(
+          onPressed: () {
+            _selectedCity = city.cityID;
+            _selectedDefaultLat = city.defaultLat;
+            _selectedDefaultLon = city.defaultLon;
+
+            _showPermissionDialog();
+          },
+          padding: EdgeInsets.zero,
+          child: Wrap(
+            children: [
+              Column(
+                children: [
+                  Padding(
+                    padding: EdgeInsets.all(16),
+                    child: Column(
+                      children: [
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          mainAxisSize: MainAxisSize.max,
                           children: [
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisSize: MainAxisSize.max,
-                              children: [
-                                Padding(
-                                    padding: EdgeInsets.zero,
-                                    child: SizedBox(
-                                        height: 64,
-                                        width: 64,
-                                        child: CircleAvatar(
-                                          backgroundColor:
-                                              Colors.purple.withOpacity(0.5),
-                                          backgroundImage:
-                                              CachedNetworkImageProvider(
-                                                  imageUrl),
-                                        ))),
-                                Expanded(
-                                    child: Padding(
-                                        padding: EdgeInsets.only(left: 16),
-                                        child: Column(
-                                          mainAxisAlignment:
-                                              MainAxisAlignment.center,
-                                          crossAxisAlignment:
-                                              CrossAxisAlignment.center,
-                                          children: [
-                                            Row(
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment.end,
-                                              children: [
-                                                Expanded(
-                                                  child: Padding(
-                                                    padding: EdgeInsets.only(
-                                                        left: 4, right: 3),
-                                                    child: Text(
-                                                      name,
-                                                      textAlign:
-                                                          TextAlign.start,
-                                                      maxLines: 1,
-                                                      overflow:
-                                                          TextOverflow.ellipsis,
-                                                      style: TextStyle(
-                                                          fontFamily:
-                                                              'LatoBold',
-                                                          fontSize: 18,
-                                                          color: Colors.white),
-                                                    ),
-                                                  ),
-                                                ),
-                                              ],
-                                            ),
-                                          ],
-                                        )))
-                              ],
-                            )
+                            _cityItemImage(city.imageUrl!),
+                            _cityItemName(city.name!),
                           ],
-                        )),
-                  ],
-                )
-              ])),
-        ));
+                        )
+                      ],
+                    ),
+                  ),
+                ],
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _cityItemImage(String imageUrl) {
+    return Padding(
+      padding: EdgeInsets.zero,
+      child: SizedBox(
+        height: 64,
+        width: 64,
+        child: CircleAvatar(
+          backgroundColor: Colors.purple.withOpacity(0.5),
+          backgroundImage: CachedNetworkImageProvider(imageUrl),
+        ),
+      ),
+    );
+  }
+
+  Widget _cityItemName(String name) {
+    return Expanded(
+      child: Padding(
+        padding: EdgeInsets.only(left: 16),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                Expanded(
+                  child: Padding(
+                    padding: EdgeInsets.only(left: 4, right: 3),
+                    child: Text(
+                      name,
+                      textAlign: TextAlign.start,
+                      maxLines: 1,
+                      overflow: TextOverflow.ellipsis,
+                      style: TextStyle(
+                          fontFamily: 'LatoBold',
+                          fontSize: 18,
+                          color: Colors.white),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }
