@@ -1,16 +1,15 @@
-import 'dart:ui';
-
-import 'package:flare_flutter/flare_controls.dart';
+import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:groovenation_flutter/constants/error.dart';
+import 'package:groovenation_flutter/constants/enums.dart';
 import 'package:groovenation_flutter/constants/strings.dart';
-import 'package:groovenation_flutter/cubit/events_cubit.dart';
+import 'package:groovenation_flutter/cubit/event/events_cubit.dart';
 import 'package:groovenation_flutter/cubit/state/events_state.dart';
 import 'package:groovenation_flutter/models/event.dart';
+import 'package:groovenation_flutter/ui/events/widgets/event_item.dart';
 import 'package:groovenation_flutter/util/alert_util.dart';
-import 'package:intl/intl.dart';
-import 'package:cached_network_image/cached_network_image.dart';
+import 'package:groovenation_flutter/widgets/custom_refresh_header.dart';
+import 'package:groovenation_flutter/widgets/top_app_bar.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class EventsHomePage extends StatefulWidget {
@@ -31,34 +30,30 @@ class EventsHomePage extends StatefulWidget {
 
 class _EventsHomePageState extends State<EventsHomePage> {
   bool _isFirstView = true;
+  final ScrollController _upcomingScrollController = new ScrollController();
+  late FavouritesEventsCubit _favouriteEventsCubit;
+
+  RefreshController _upcomingRefreshController =
+      RefreshController(initialRefresh: true);
+  final _favouritesRefreshController = RefreshController(initialRefresh: false);
+
 
   runBuild() {
     if (_isFirstView) {
       int pg = widget.page;
       print("Running Build: $pg");
       _isFirstView = false;
-
-      // final FavouritesEventsCubit favouriteEventsCubit =
-      //     BlocProvider.of<FavouritesEventsCubit>(context);
-
-      // favouriteEventsCubit.getEvents(0);
     }
   }
 
-  final FlareControls flareControls = FlareControls();
-  final ScrollController _upcomingScrollController = new ScrollController();
 
-  RefreshController _upcomingRefreshController =
-      RefreshController(initialRefresh: true);
-  final _favouritesRefreshController = RefreshController(initialRefresh: false);
 
-  void setListScrollListener(
-      ScrollController controller, RefreshController refreshController) {
+  void _setListScrollListener(ScrollController controller) {
     controller.addListener(() {
       if (controller.position.pixels >=
           (controller.position.maxScrollExtent - 456)) {
-        if (refreshController.footerStatus == LoadStatus.idle) {
-          if(mounted) refreshController.requestLoading(needMove: false);
+        if (_upcomingRefreshController.footerStatus == LoadStatus.idle) {
+          _upcomingRefreshController.requestLoading(needMove: false);
         }
       }
     });
@@ -68,87 +63,45 @@ class _EventsHomePageState extends State<EventsHomePage> {
   void initState() {
     super.initState();
 
-      final FavouritesEventsCubit favouriteEventsCubit =
-        BlocProvider.of<FavouritesEventsCubit>(context);
+    _favouriteEventsCubit = BlocProvider.of<FavouritesEventsCubit>(context);
+    _favouriteEventsCubit.getEvents(0);
 
-      favouriteEventsCubit.getEvents(0);
-
-      WidgetsBinding.instance
-        .addPostFrameCallback((_) => setListScrollListener(
-        _upcomingScrollController, _upcomingRefreshController));
+    _setListScrollListener(_upcomingScrollController);
   }
-
-  openSearchPage() {
-    Navigator.pushNamed(context, '/search');
-  }
-
-  Column topAppBar() => Column(children: [
-        Padding(
-          padding: EdgeInsets.only(top: 16, left: 16, right: 16, bottom: 16),
-          child: FlatButton(
-              padding: EdgeInsets.zero,
-              onPressed: () {
-                openSearchPage();
-              },
-              child: Container(
-                  width: double.infinity,
-                  height: 56,
-                  decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(10.0),
-                      color: Colors.white.withOpacity(0.2)),
-                  child: Stack(
-                    children: [
-                      Align(
-                        alignment: Alignment.centerLeft,
-                        child: Padding(
-                          padding: EdgeInsets.only(left: 24),
-                          child: Text(
-                            "Search",
-                            style: TextStyle(
-                                color: Colors.white.withOpacity(0.5),
-                                fontFamily: 'Lato',
-                                fontSize: 17),
-                          ),
-                        ),
-                      ),
-                      Align(
-                        alignment: Alignment.centerRight,
-                        child: Padding(
-                            padding: EdgeInsets.only(right: 24),
-                            child: Icon(
-                              Icons.search,
-                              size: 28,
-                              color: Colors.white.withOpacity(0.5),
-                            )),
-                      ),
-                    ],
-                  ))),
-        ),
-        TabBar(
-          tabs: [
-            Tab(
-              icon: Icon(Icons.whatshot),
-              text: "Upcoming",
-            ),
-            Tab(
-              icon: Icon(Icons.favorite),
-              text: "Favourites",
-            ),
-          ],
-        )
-      ]);
 
   @override
   void dispose() {
+    super.dispose();
     _upcomingScrollController.dispose();
     _upcomingRefreshController.dispose();
+  }
 
-    // _favouritesScrollController.dispose();
-    // _favouritesRefreshController.dispose();
-    super.dispose();
+  void _checkEventError(EventsState state) {
+    if (state is EventsErrorState) {
+      switch (state.error) {
+        case AppError.NETWORK_ERROR:
+          alertUtil.sendAlert(
+              BASIC_ERROR_TITLE, NETWORK_ERROR_PROMPT, Colors.red, Icons.error);
+          break;
+        default:
+          alertUtil.sendAlert(
+              BASIC_ERROR_TITLE, UNKNOWN_ERROR_PROMPT, Colors.red, Icons.error);
+          break;
+      }
+    }
   }
 
   bool _isInitialFavouriteLoad = true;
+  final List<Tab> _tabs = [
+    Tab(
+      icon: Icon(Icons.whatshot),
+      text: "Upcoming",
+    ),
+    Tab(
+      icon: Icon(Icons.favorite),
+      text: "Favourites",
+    ),
+  ];
 
   @override
   Widget build(BuildContext context) {
@@ -156,91 +109,103 @@ class _EventsHomePageState extends State<EventsHomePage> {
     if (pg == 1 && _isFirstView) runBuild();
 
     return BlocConsumer<FavouritesEventsCubit, EventsState>(
-        listener: (context, favouriteEventsState) {
-      if (favouriteEventsState is EventsLoadedState && _isFirstView == false) {
-        final UpcomingEventsCubit upcomingEventsCubit =
-            BlocProvider.of<UpcomingEventsCubit>(context);
-        if (_isInitialFavouriteLoad) {
-          _isInitialFavouriteLoad = false;
-          upcomingEventsCubit.getEvents(upcomingPage);
+      listener: _favouritesBlocListener,
+      builder: (context, favouriteEventsState) {
+        if (favouriteEventsState is EventsLoadedState) {
+          favouriteEvents = favouriteEventsState.events;
         }
-      }
-    }, builder: (context, favouriteEventsState) {
-      if (favouriteEventsState is EventsLoadedState) {
-        favouriteEvents = favouriteEventsState.events;
-      }
 
-      return SafeArea(
+        return SafeArea(
           child: DefaultTabController(
-              length: 2,
-              child: Column(
-                children: [
-                  topAppBar(),
-                  Expanded(
-                    child: TabBarView(
-                      children: [
-                        upcomingList(),
-                        SmartRefresher(
-                          controller: _favouritesRefreshController,
-                          header: WaterDropMaterialHeader(),
-                          enablePullDown: false,
-                          enablePullUp: false,
-                          child: favouritesList(),
-                        ),
-                      ],
-                    ),
-                  )
-                ],
-              )));
-    });
+            length: 2,
+            child: Column(
+              children: [
+                TopAppBar(tabs: _tabs),
+                Expanded(
+                  child: TabBarView(
+                    children: [
+                      _upcomingList(),
+                      _favouritesList(),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
   }
 
-  List<Event> upcomingEvents = [];
+  //Upcoming Events:
+
+  List<Event>? upcomingEvents = [];
   int upcomingPage = 0;
 
-  Widget upcomingList() {
-    return BlocConsumer<UpcomingEventsCubit, EventsState>(
-        listener: (context, state) {
-      if (state is EventsLoadedState) {
-        // setState(() {
-        //         upcomingPage++;
-        //       });
-        if (_upcomingRefreshController.isRefresh) {
-          _upcomingScrollController.jumpTo(0.0);
-          _upcomingRefreshController.refreshCompleted();
+  void _refreshUpcoming() {
+    if (!_isFavouritesRequestSuccess()) return;
+
+    final UpcomingEventsCubit upcomingEventsCubit =
+        BlocProvider.of<UpcomingEventsCubit>(context);
+
+    if ((upcomingEventsCubit.state is EventsLoadedState ||
+            upcomingEventsCubit.state is EventsErrorState) &&
+        !_isFirstView) {
+      setState(() {
+        upcomingPage = 0;
+      });
+
+      _upcomingRefreshController.loadComplete();
+
+      upcomingEventsCubit.getEvents(0);
+    }
+  }
+
+  void _loadMoreUpcoming() {
+    if (upcomingEvents!.length == 0) {
+      _upcomingRefreshController.loadComplete();
+      return;
+    }
+
+    final UpcomingEventsCubit upcomingEventsCubit =
+        BlocProvider.of<UpcomingEventsCubit>(context);
+
+    setState(() {
+      upcomingPage++;
+    });
+
+    upcomingEventsCubit.getEvents(upcomingPage);
+  }
+
+  void _upcomingBlocListener(context, state) {
+    if (state is EventsLoadedState) {
+      if (_upcomingRefreshController.isRefresh) {
+        _upcomingScrollController.jumpTo(0.0);
+        _upcomingRefreshController.refreshCompleted();
+        _upcomingRefreshController.loadComplete();
+
+        _upcomingRefreshController = RefreshController(initialRefresh: false);
+      } else if (_upcomingRefreshController.isLoading) {
+        if (state.hasReachedMax!)
+          _upcomingRefreshController.loadNoData();
+        else
           _upcomingRefreshController.loadComplete();
-
-          _upcomingRefreshController = RefreshController(initialRefresh: false);
-        } else if (_upcomingRefreshController.isLoading) {
-          if (state.hasReachedMax)
-            _upcomingRefreshController.loadNoData();
-          else
-            _upcomingRefreshController.loadComplete();
-        }
       }
-    }, builder: (context, upcomingEventsState) {
-      if (upcomingEventsState is EventsLoadedState)
-        upcomingEvents = upcomingEventsState.events;
+    }
+  }
 
-      if (upcomingEventsState is EventsErrorState) {
-        switch (upcomingEventsState.error) {
-          case Error.NETWORK_ERROR:
-            alertUtil.sendAlert(BASIC_ERROR_TITLE, NETWORK_ERROR_PROMPT,
-                Colors.red, Icons.error);
-            break;
-          default:
-            alertUtil.sendAlert(BASIC_ERROR_TITLE, UNKNOWN_ERROR_PROMPT,
-                Colors.red, Icons.error);
-            break;
-        }
-      }
+  Widget _upcomingList() {
+    return BlocConsumer<UpcomingEventsCubit, EventsState>(
+      listener: _upcomingBlocListener,
+      builder: (context, upcomingEventsState) {
+        if (upcomingEventsState is EventsLoadedState)
+          upcomingEvents = upcomingEventsState.events;
 
-      final FavouritesEventsCubit favouritesEventsCubit =
-          BlocProvider.of<FavouritesEventsCubit>(context);
+        _checkEventError(upcomingEventsState);
 
-      return SmartRefresher(
+        return SmartRefresher(
           controller: _upcomingRefreshController,
-          header: WaterDropMaterialHeader(),
+          header: CustomMaterialClassicHeader(),
           footer: ClassicFooter(
             textStyle: TextStyle(
                 color: Colors.white.withOpacity(0.5),
@@ -249,222 +214,76 @@ class _EventsHomePageState extends State<EventsHomePage> {
             noDataText: "You've reached the end of the line",
             failedText: "Something Went Wrong",
           ),
-          onRefresh: () {
-            print("favsuc:" + checkFavouritesSuccess().toString());
-
-            if (!checkFavouritesSuccess()) return;
-
-            final UpcomingEventsCubit upcomingEventsCubit =
-                BlocProvider.of<UpcomingEventsCubit>(context);
-
-            if ((upcomingEventsCubit.state is EventsLoadedState ||
-                    upcomingEventsCubit.state is EventsErrorState) &&
-                !_isFirstView) {
-              setState(() {
-                upcomingPage = 0;
-              });
-
-              _upcomingRefreshController.loadComplete();
-
-              upcomingEventsCubit.getEvents(0);
-            }
-          },
-          onLoading: () {
-            if (upcomingEvents.length == 0) {
-              _upcomingRefreshController.loadComplete();
-              return;
-            }
-
-            final UpcomingEventsCubit upcomingEventsCubit =
-                BlocProvider.of<UpcomingEventsCubit>(context);
-
-            setState(() {
-              upcomingPage++;
-            });
-
-            upcomingEventsCubit.getEvents(upcomingPage);
-          },
+          onRefresh: _refreshUpcoming,
+          onLoading: _loadMoreUpcoming,
           enablePullUp: true,
           child: ListView.builder(
-              controller: _upcomingScrollController,
-              padding: EdgeInsets.only(top: 16, bottom: 0, left: 16, right: 16),
-              physics: const BouncingScrollPhysics(
-                  parent: AlwaysScrollableScrollPhysics()),
-              itemCount: upcomingEvents.length,
-              itemBuilder: (context, index) {
-                return Padding(
-                    padding: EdgeInsets.only(bottom: 12),
-                    child: eventItem(
-                        context,
-                        upcomingEvents[index],
-                        favouritesEventsCubit
-                            .checkEventExists(upcomingEvents[index].eventID),
-                        index));
-              }));
-    });
+            controller: _upcomingScrollController,
+            padding: EdgeInsets.only(top: 16, bottom: 0, left: 16, right: 16),
+            physics: const BouncingScrollPhysics(
+                parent: AlwaysScrollableScrollPhysics()),
+            itemCount: upcomingEvents!.length,
+            itemBuilder: (context, index) {
+              return Padding(
+                padding: EdgeInsets.only(bottom: 12),
+                child: EventItem(
+                  event: upcomingEvents![index],
+                  isFavourite: _favouriteEventsCubit
+                      .checkEventExists(upcomingEvents![index].eventID),
+                ),
+              );
+            },
+          ),
+        );
+      },
+    );
   }
 
-  bool checkFavouritesSuccess() {
-    final FavouritesEventsCubit favouritesEventsCubit =
-        BlocProvider.of<FavouritesEventsCubit>(context);
-    if (favouritesEventsCubit.state is EventsErrorState) {
-      favouritesEventsCubit.getEvents(0);
+  //Favourite Events:
+
+  List<Event>? favouriteEvents = [];
+
+  void _favouritesBlocListener(context, favouriteEventsState) {
+    if (favouriteEventsState is EventsLoadedState && _isFirstView == false) {
+      final UpcomingEventsCubit upcomingEventsCubit =
+          BlocProvider.of<UpcomingEventsCubit>(context);
+      if (_isInitialFavouriteLoad) {
+        _isInitialFavouriteLoad = false;
+        upcomingEventsCubit.getEvents(upcomingPage);
+      }
+    }
+  }
+
+  bool _isFavouritesRequestSuccess() {
+    if (_favouriteEventsCubit.state is EventsErrorState) {
+      _favouriteEventsCubit.getEvents(0);
       return false;
     }
-    if (favouritesEventsCubit.state is EventsLoadingState) {
+    if (_favouriteEventsCubit.state is EventsLoadingState) {
       return false;
     }
     return true;
   }
 
-  List<Event> favouriteEvents = [];
-
-  Widget favouritesList() {
-    return ListView.builder(
+  Widget _favouritesList() {
+    return SmartRefresher(
+      controller: _favouritesRefreshController,
+      header: CustomMaterialClassicHeader(),
+      enablePullDown: false,
+      enablePullUp: false,
+      child: ListView.builder(
         physics: const BouncingScrollPhysics(
-            parent: AlwaysScrollableScrollPhysics()),
+          parent: AlwaysScrollableScrollPhysics(),
+        ),
         padding: EdgeInsets.only(top: 16, bottom: 0, left: 16, right: 16),
-        itemCount: favouriteEvents.length,
+        itemCount: favouriteEvents!.length,
         itemBuilder: (context, index) {
           return Padding(
-              padding: EdgeInsets.only(bottom: 12),
-              child: eventItem(context, favouriteEvents[index], true, index));
-        });
-  }
-
-  Widget eventItem(
-      BuildContext context, Event event, bool isFavourite, int index) {
-    return Card(
-      elevation: 6,
-      color: Colors.deepPurple,
-      clipBehavior: Clip.antiAliasWithSaveLayer,
-      shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(Radius.circular(12.0))),
-      child: FlatButton(
-          onPressed: () {
-            Navigator.pushNamed(context, '/event', arguments: event);
-          },
-          padding: EdgeInsets.zero,
-          child: Wrap(children: [
-            Column(
-              children: [
-                Row(mainAxisSize: MainAxisSize.max, children: [
-                  Expanded(
-                    child: Container(
-                      height: 256,
-                      decoration: BoxDecoration(
-                        image: DecorationImage(
-                            image: CachedNetworkImageProvider(event.imageUrl),
-                            fit: BoxFit.cover),
-                      ),
-                      child: Align(
-                        alignment: Alignment.topRight,
-                        child: Padding(
-                            padding: EdgeInsets.only(top: 16, right: 16),
-                            child: Container(
-                              height: 48,
-                              width: 48,
-                              decoration: BoxDecoration(
-                                  color: Colors.deepPurple,
-                                  borderRadius: BorderRadius.circular(9)),
-                              child: FlatButton(
-                                  padding: EdgeInsets.zero,
-                                  onPressed: () {
-                                    final FavouritesEventsCubit
-                                        favouritesEventsCubit =
-                                        BlocProvider.of<FavouritesEventsCubit>(
-                                            context);
-                                    if (isFavourite) {
-                                      favouritesEventsCubit.removeEvent(event);
-                                    } else
-                                      favouritesEventsCubit.addEvent(event);
-                                  },
-                                  child: Icon(
-                                    isFavourite
-                                        ? Icons.favorite
-                                        : Icons.favorite_border,
-                                    color: Colors.white,
-                                    size: 28,
-                                  )),
-                            )),
-                      ),
-                    ),
-                  ),
-                ]),
-                Padding(
-                    padding: EdgeInsets.all(16),
-                    child: Column(
-                      children: [
-                        Row(
-                          mainAxisSize: MainAxisSize.max,
-                          children: [
-                            SizedBox(
-                                height: 76,
-                                width: 76,
-                                child: CircleAvatar(
-                                    backgroundColor: Colors.white,
-                                    child: Container(
-                                      child: Center(
-                                          child: Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        children: [
-                                          Text(
-                                            DateFormat.MMM()
-                                                .format(event.eventStartDate),
-                                            style: TextStyle(
-                                              fontFamily: 'LatoBold',
-                                              fontSize: 18,
-                                            ),
-                                          ),
-                                          Text(
-                                            DateFormat.d()
-                                                .format(event.eventStartDate),
-                                            style: TextStyle(
-                                              fontFamily: 'LatoBold',
-                                              fontSize: 18,
-                                            ),
-                                          ),
-                                        ],
-                                      )),
-                                    ))),
-                            Container(
-                              padding: EdgeInsets.only(left: 20, top: 0),
-                              child: Column(
-                                mainAxisAlignment: MainAxisAlignment.start,
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  Text(
-                                    event.title,
-                                    textAlign: TextAlign.start,
-                                    style: TextStyle(
-                                        fontFamily: 'LatoBold',
-                                        fontSize: 22,
-                                        color: Colors.white),
-                                  ),
-                                  Padding(
-                                      padding: EdgeInsets.only(top: 6),
-                                      child: Text(
-                                        event.clubName,
-                                        textAlign: TextAlign.start,
-                                        style: TextStyle(
-                                            fontFamily: 'Lato',
-                                            fontSize: 18,
-                                            color:
-                                                Colors.white.withOpacity(0.4)),
-                                      )),
-                                ],
-                              ),
-                            ),
-                          ],
-                        )
-                      ],
-                    )),
-              ],
-            )
-          ])),
+            padding: EdgeInsets.only(bottom: 12),
+            child: EventItem(event: favouriteEvents![index], isFavourite: true),
+          );
+        },
+      ),
     );
   }
 }
